function_calling_project/requirements.txt
openai>=0.27.0
python-dotenv>=1.0.0
